+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "apimock"
-version = "2.1.0"
+version = "2.2.0"
 edition = "2021"
 authors = ["nabbisen <nabbisen@scqr.net>"]
 license = "ISC"
@@ -25,3 +25,6 @@ serde = "1"
 serde_json = "1"
 json5 = "^0.4"
 console = "^0.15"
+
+[dev-dependencies]
+hyper = { version = "0.14", features = ["client"] }
